requirements.txt
annotated-types==0.7.0
anyio==4.8.0
click==8.1.8
colorama==0.4.6
dnspython==2.7.0
fastapi==0.115.11
h11==0.14.0
idna==3.10
Jinja2==3.1.6
MarkupSafe==3.0.2
motor==3.7.0
pydantic==2.10.6
pydantic_core==2.27.2
pymongo==4.11.2
sniffio==1.3.1
starlette==0.46.1
typing_extensions==4.12.2
uvicorn==0.34.0
websockets==15.0.1
python-dateutil==2.8.2
pytz